feat: add eslint in project
