+++ .eslintrc.js
@@ -1,0 +1,18 @@
+module.exports = {
+    "env": {
+        "browser": true,
+        "es2021": true
+    },
+    "extends": "eslint:recommended",
+    "overrides": [
+    ],
+    "parserOptions": {
+        "ecmaVersion": "latest",
+        "sourceType": "module"
+    },
+    "rules": {
+        'no-undef': 'off',
+        "no-useless-escape": "off"
+    },
+    "plugins": ['node']
+}
