+++ package.json
@@ -1,7 +1,8 @@
 {
   "scripts": {
     "start": "webpack-dev-server --open --config webpack.config.dev.js",
-    "build": "webpack --config webpack.config.prod.js"
+    "build": "webpack --config webpack.config.prod.js",
+    "lint": "eslint ."
   },
   "dependencies": {
     "copy-webpack-plugin": "^11.0.0",
@@ -18,6 +19,8 @@
     "webpack-5": "^0.0.0"
   },
   "devDependencies": {
+    "eslint": "^8.39.0",
+    "eslint-plugin-node": "^11.1.0",
     "file-loader": "^6.2.0",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.1",
